general: start implementing the editor widget.

+++ src/vec.zig
@@ -1,4 +1,4 @@
 // TODO(remy): comment me
 // TODO(remy): struct name isn't ideal
 // TODO(remy): (a,b) isn't optimal
-pub const Vec2i = struct { a: i32, b: i32 };
+pub const Vec2i = struct { a: i64, b: i64 };
